partner choice with elitist selection
Former-commit-id: afe6415b0b9af5b8f64fb6ec3abbfb3245c7d1c6

+++ roborobo3/include/core/RoboroboMain/roborobo.h
@@ -121,7 +121,7 @@ extern bool gStepByStep;
 extern bool gInspectorMode;			// defines if manual control is robot-centered or "god"-centered (ie. envt viewpoint)
 extern bool gInspectorAgent;		// defines if there *actually* is an inspector agent (note: if using batch mode for, e.g. evolution, no inspector agent is needed)
 
-extern int gMaxIt;
+extern long long gMaxIt;
 
 extern int gNbOfLandmarks;
 extern std::vector<LandmarkObject*> gLandmarks; // list of inanimate objects (landmarks).
